proc v1.3.5.1 Rev2
v1.3.5.1-Rev2 is a quick revision update and fix for v1.3.5-RC2 stuck on
block 367919. This is an important update for anyone running v1.3.5-RC2
or stuck on block 367919. The Protocol version has been bumped to 75530
and also some few changes to the GUI wallet. The Yearly Blockcount has
been put back to 525600 which is the Block Number on which the reward
will halve to 10 PROC/ block. Please update to get your wallet moving.