proc v1.39.2.3 Mandatory Update
This Quick update (v1.39.2.3) is to fix a Minor Issue in v1.3.9.2.2 LR3
that cause the Blockchain to be stuck when doing a Sync from scratch or
Clean Sync.

IMPORTANT: The Forks are still active and we have an additional Fork at
Block# 639000
to Enforce the Blocktime. It is MANDATORY that everybody be on this
Latest Release.
-FORK1 #649000: New Coinbase maturity set to 220 confirmations
-FORK2 #649001: New Min Stake Age set to 2hrs
-FORK3 #649002: New Min TxFee set to 0.001
-FORK4: #639000: New Blocktime set to 240s

-Fixed some Minor issues to allow the Blockchain to Sync from scratch
-Added new Blocktime fork - Activation block# 639000
-Added "PrivateSend" tab to the Navigation Panel